Added logic to check_fdecl to ensure that:
1. The Analyzer throws an error if the param has been constrained to
one type, and the function is called on another type.
2. The Analyzer throws an error if the function’s parameter types have
been constrained to one type, but the function’s parameters are
constrained to another type.
3. The analyzer checks param types against the function type’s
parameter types, and tries to constrain them if possible, and
vice-versa (tries to constrain function parameter types if the function
type’s parameter type’s themselves have been constrained).
3. The analyzer reevaluates the function’s return type to see if any of
the steps above have constrained the return type’s expression. If so,
constrain the return type.