Fixed 4 bugs in Analyzer
1. Improper error was being called called on return type mismatch
2. Fixed constrain_ew constrain error checks - though we never ran into
any issues w/ these, the way we were doing it (old_type <> Unconst)
might have thrown an error if we tried to constrain a List(Unconst) to
a List(const).
3. MOST IMPORTANT: any function that returns a List(Any) will, upon
invocation, return a List(Unconst) - b4 we were returning List(Any)
which could have led to some serious issues down the line. The only
reason we were not having issues before is actually because of the bug
in constrain_ew, which was allowing us to constrain List(Any) to
List(Num) or any other List(const).
4. Fixed bug in check_fdecl where an error (an improper error, see 1)
was being thrown if return type was List(Any).